requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.0
tavily-python==0.3.3
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0